+++ src/components/DetailsSelectBox/DetailsSelectBox.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import SelectBox from "./SelectBox";
+import DetailsSelectBox from "./DetailsSelectBox";
 
 
 let options = [];
@@ -18,7 +18,7 @@ describe("testing the creation of 1 select drop down and handling change", () =>
   test("renders 1 drop down with an initial value and required text ", () => {
 
     const { container } = render(
-      <SelectBox
+      <DetailsSelectBox
         caption={"test caption"}
         required={true}
         initialSelection={1}
@@ -33,7 +33,7 @@ describe("testing the creation of 1 select drop down and handling change", () =>
   test("renders 1 drop down with no initial value and no required text  ", () => {
 
     const { container } = render(
-      <SelectBox
+      <DetailsSelectBox
         caption={"test caption"}
         initialSelection={null}
         required={false}
