+++ src/components/SelectBox/SelectBox.js
@@ -3,7 +3,7 @@ import {
   getActiveConfigurations,
   getInActiveConfigurations,
 } from "../../utils/selectors/monitoringConfigurations";
-
+import { Label, Dropdown, FormGroup } from "@trussworks/react-uswds";
 import "./SelectBox.scss";
 
 const SelectBox = ({
@@ -13,6 +13,8 @@ const SelectBox = ({
   selectionHandler,
   showInactive = false,
   initialSelection,
+  viewOnly,
+  required,
 }) => {
   function getIndex(name) {
     return options.findIndex((obj) => obj[selectKey] === name);
@@ -23,9 +25,7 @@ const SelectBox = ({
 
   const handleChange = (val) => {
     setSelectionState(getIndex(val.target.value));
-    selectionHandler(getIndex(val.target.value));
   };
-
   const populateOptions = (optionsList) => {
     return optionsList.map((info, index) => {
       return (
@@ -39,38 +39,48 @@ const SelectBox = ({
     selectionHandler(initialSelection ? initialSelection : 0);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+
+  useEffect(() => {
+    selectionHandler(selectionState);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [selectionState]);
+
   return (
     <div>
-      <div className="mpSelect">
-        {caption}
-        {": "} <br />
-        <select
-          role="select"
-          data-testid="select"
-          id={caption}
-          onChange={(e) => handleChange(e)}
-          value={
-            options[selectionState] !== undefined
-              ? options[selectionState][selectKey]
-              : options[0][selectKey]
-          }
-        >
-          {showInactive &&
-            caption === "Configurations" &&
-            getActiveConfigurations(options).length > 0 && (
-              <optgroup label="Active" role="optGroup">
-                {populateOptions(getActiveConfigurations(options))}
-              </optgroup>
-            )}
-          {showInactive &&
-            caption === "Configurations" &&
-            getInActiveConfigurations(options).length > 0 && (
-              <optgroup label="Inactive" role="optGroup">
-                {populateOptions(getInActiveConfigurations(options))}
-              </optgroup>
-            )}
-          {showInactive === false && populateOptions(options)}
-        </select>
+      <div className="">
+        <FormGroup className="formLabeling">
+          <Label>{caption}</Label>
+          <Dropdown
+            aria-label={caption + " dropdown"}
+            name={caption}
+            // weird bug without this
+            defaultValue={
+              options[selectionState] !== undefined
+                ? options[selectionState][selectKey]
+                : options[0][selectKey]
+            }
+            disabled={viewOnly}
+            role="select"
+            id={caption + initialSelection}
+            onChange={(e) => handleChange(e)}
+          >
+            {showInactive &&
+              caption === "Configurations" &&
+              getActiveConfigurations(options).length > 0 && (
+                <optgroup label="Active" role="optGroup">
+                  {populateOptions(getActiveConfigurations(options))}
+                </optgroup>
+              )}
+            {showInactive &&
+              caption === "Configurations" &&
+              getInActiveConfigurations(options).length > 0 && (
+                <optgroup label="Inactive" role="optGroup">
+                  {populateOptions(getInActiveConfigurations(options))}
+                </optgroup>
+              )}
+            {showInactive === false && populateOptions(options)}
+          </Dropdown>
+        </FormGroup>
       </div>
     </div>
   );
